Update to pe_code_manager_webhook v1.0.0
Prior to this commit, we were using a pre-release version of the
pe_code_manager_webhook module in order to test that simply
re-namespacing the module form a profile to a module worked as
expected.

After this commit, we use the official 1.0.0 relase of the module
that includes a few changes that make it more versatile.