Clean up some stale files
Prior to this commit, these two files were laying around but not used.
The code manager template was missed when the code was moved into
a module and the function was added a while ago but then never
put to use.

After this commit, the codebase is a little lighter without these
stale files.