Move gitlab profile out to its own 'profile' module
Prior to this commit, there was a gitlab profile in the site dir
of this control-repo.

After this commit, the profile has been moved to it's own repo
where it can be used more generally and not tightly coupled to this
repository.

As a result, the gitlab role has been removed from this repository
as well.