Reorganize the README and improve the documentation
This commit moves the "where did all the previous code go" section to
the bottom as it's been a while since that change was made. Nowadays,
people new to Control Repositories will find this and won't understand
the reference as they never knew about previous versions.

Now the README starts right away with information on what this project
is and how to use it.

Also cleaned up some of the Markdown syntax to make it easier to read.