Rename site directory to site-modules
Prior to this commit, we placed modules local to a users installation
in the `site` directory.  This was just a convention and the name
`site` doesn't clearly convey what it is for.

After this commit, we place modules local to a users installation in
the `site-modules` directory.  This makes it more clear to users
that this is a directory that modules go i.  When users start
with bolt they won't even know what a control-repo is and
renaming site to site-modules gives them a better idea of why
they should put their modules with tasks in them.  Also see:

https://tickets.puppetlabs.com/browse/BOLT-1108